tweaks to enable resources from hub solutions to upload
