+++ packages/common/src/index.ts
@@ -33,3 +33,4 @@ export * from "./restHelpersGet";
 export * from "./sharing";
 export * from "./templatization";
 export * from "./migrator";
+export * from "./resources";
