+++ packages/common/src/resourceHelpers.ts
@@ -59,20 +59,20 @@ import {
   IFileMimeType,
   IItemTemplate,
   IItemUpdate,
-  IMimeTypes,
   ISourceFileCopyPath,
   IUpdateItemResponse,
   UserSession
 } from "./interfaces";
 import { new_File } from "./polyfills";
 import {
-  addItemResource,
   updateGroup,
   updateItem,
   updateItemInfo,
   updateItemResource
 } from "@esri/arcgis-rest-portal";
-import { ArcGISAuthError } from "@esri/arcgis-rest-request";
+import { addResourceFromBlob } from "./resources/add-resource-from-blob";
+import { copyResource } from "./resources/copy-resource";
+
 import { updateItem as helpersUpdateItem } from "./restHelpers";
 import { getBlob, getBlobAsFile, getItemResources } from "./restHelpersGet";
 
@@ -102,39 +102,6 @@ export function addMetadataFromBlob(
     authentication: authentication
   };
   return updateItem(updateOptions);
-}
-
-export function addResourceFromBlob(
-  blob: any,
-  itemId: string,
-  folder: string,
-  filename: string,
-  authentication: UserSession
-): Promise<any> {
-  // Check that the filename has an extension because it is required by the addResources call
-  if (filename && filename.indexOf(".") < 0) {
-    return new Promise((resolve, reject) => {
-      reject(
-        new ArcGISAuthError(
-          "Filename must have an extension indicating its type"
-        )
-      );
-    });
-  }
-
-  const addRsrcOptions = {
-    id: itemId,
-    resource: blob,
-    name: filename,
-    authentication: authentication,
-    params: {}
-  };
-  if (folder) {
-    addRsrcOptions.params = {
-      resourcesPrefix: folder
-    };
-  }
-  return addItemResource(addRsrcOptions);
 }
 
 export function addThumbnailFromBlob(
@@ -247,25 +214,28 @@ export function copyFilesFromStorageItem(
   destinationItemId: string,
   destinationAuthentication: UserSession,
   isGroup: boolean = false,
-  template: any
+  template: any = {}
 ): Promise<boolean> {
   // TODO: This is only used in deployer, so move there
-
   // changed to allow the template to be passed in
   // because Hub templates need to swap out the templateId
   // in the reseource filename
   const mimeTypes = template.properties || null;
 
   // remove the template.itemId from the fileName in the filePaths
-  filePaths = filePaths.map(fp => {
-    if (fp.filename.indexOf(template.itemId) === 0 && fp.folder === "") {
-      fp.filename = fp.filename.replace(`${template.itemId}-`, "");
-    }
-    return fp;
-  });
+  if (template.itemId) {
+    filePaths = filePaths.map(fp => {
+      if (fp.filename.indexOf(template.itemId) === 0 && fp.folder === "") {
+        fp.filename = fp.filename.replace(`${template.itemId}-`, "");
+      }
+      return fp;
+    });
+  }
 
   return new Promise<boolean>((resolve, reject) => {
     // Introduce a lag because AGO update appears to choke with rapid subsequent calls
+    // Note: This is not actually delaying. The map returns an array of promises
+    // all of which will start firing in `lagMs` milliseconds
     const msLag = 1000;
 
     const awaitAllItems = filePaths.map(filePath => {
@@ -470,65 +440,6 @@ export function copyMetadata(
 }
 
 /**
- * Copies a resource from a URL to an item.
- *
- * @param source.url URL to source resource
- * @param source.authentication Credentials for the request to source
- * @param destination.itemId Id of item to receive copy of resource/metadata/thumbnail
- * @param destination.folderName Folder in destination for resource/metadata/thumbnail; defaults to top level
- * @param destination.filename Filename in destination for resource/metadata/thumbnail
- * @param destination.authentication Credentials for the request to destination
- * @return A promise which resolves to the filename under which the resource/metadata/thumbnail is stored
- */
-export function copyResource(
-  source: {
-    url: string;
-    authentication: UserSession;
-  },
-  destination: {
-    itemId: string;
-    folder: string;
-    filename: string;
-    authentication: UserSession;
-  }
-): Promise<any> {
-  return new Promise<any>((resolve, reject) => {
-    getBlob(source.url, source.authentication).then(
-      async blob => {
-        if (
-          blob.type.startsWith("text/plain") ||
-          blob.type === "application/json"
-        ) {
-          try {
-            const text = await new Response(blob).text();
-            const json = JSON.parse(text);
-            if (json.error) {
-              reject(); // unable to get resource
-              return;
-            }
-          } catch (Ignore) {
-            reject(); // unable to get resource
-            return;
-          }
-        }
-
-        addResourceFromBlob(
-          blob,
-          destination.itemId,
-          destination.folder,
-          destination.filename,
-          destination.authentication
-        ).then(
-          resolve,
-          e => reject(fail(e)) // unable to add resource
-        );
-      },
-      e => reject(fail(e)) // unable to get resource
-    );
-  });
-}
-
-/**
  * Generates the full URL and storage folder/filename for storing the thumbnail of a group.
  *
  * @param portalSharingUrl Server/sharing
@@ -636,6 +547,8 @@ export function generateResourceFilenameFromStorage(
     const folderStart = folder.indexOf("_");
     if (folderStart > 0) {
       folder = folder.substr(folderStart + 1);
+    } else {
+      folder = "";
     }
   }
